Extract the logic for use_path to its own module
Rework the UseRename case to preserve the original Ident and the renamed
one. This way we could restore it later when building the new UseTree

Now the replacement is more inconvenient:
```rust
use crate::one::This as That;
```
=>
```rust
replace!(
 crate::one::This__synca_as_That => crate::two::This__synca_as_That
)
```
