+++ synca/src/fold/fold.rs
@@ -8,8 +8,6 @@ use syn::{
 
 use crate::SyncAFoldAttributes;
 
-const GLOB_IDENTIFIER: &'static str = "__synca_glob_identifier";
-
 #[derive(Debug, PartialEq)]
 pub struct SyncAFold {
   pub module_name: String,
@@ -60,166 +58,7 @@ impl Fold for SyncAFold {
       return fold::fold_use_path(self, use_path);
     }
 
-    // Step 1: Flatten use_path into Vec<syn::Type::Path>
-    fn flatten_use_path(
-      use_tree: &syn::UseTree,
-      prefix: Vec<syn::PathSegment>,
-      out: &mut Vec<syn::Type>,
-    ) {
-      match use_tree {
-        syn::UseTree::Path(use_path) => {
-          let mut new_prefix = prefix.clone();
-          new_prefix.push(syn::PathSegment {
-            ident: use_path.ident.clone(),
-            arguments: syn::PathArguments::None,
-          });
-          flatten_use_path(&use_path.tree, new_prefix, out);
-        }
-        syn::UseTree::Name(use_name) => {
-          let mut new_prefix = prefix.clone();
-          new_prefix.push(syn::PathSegment {
-            ident: use_name.ident.clone(),
-            arguments: syn::PathArguments::None,
-          });
-          out.push(syn::Type::Path(syn::TypePath {
-            qself: None,
-            path: syn::Path {
-              leading_colon: None,
-              segments: new_prefix.into_iter().collect(),
-            },
-          }));
-        }
-        syn::UseTree::Group(use_group) => {
-          for item in &use_group.items {
-            flatten_use_path(item, prefix.clone(), out);
-          }
-        }
-        syn::UseTree::Glob(_use_glob) => {
-          let mut new_prefix = prefix.clone();
-          new_prefix.push(syn::PathSegment {
-            ident: syn::Ident::new(GLOB_IDENTIFIER, proc_macro2::Span::call_site()),
-            arguments: syn::PathArguments::None,
-          });
-          out.push(syn::Type::Path(syn::TypePath {
-            qself: None,
-            path: syn::Path {
-              leading_colon: None,
-              segments: new_prefix.into_iter().collect(),
-            },
-          }));
-        }
-        syn::UseTree::Rename(use_rename) => {
-          let mut new_prefix = prefix.clone();
-          new_prefix.push(syn::PathSegment {
-            ident: use_rename.rename.clone(),
-            arguments: syn::PathArguments::None,
-          });
-          out.push(syn::Type::Path(syn::TypePath {
-            qself: None,
-            path: syn::Path {
-              leading_colon: None,
-              segments: new_prefix.into_iter().collect(),
-            },
-          }));
-        }
-      }
-    }
-
-    let mut flat_paths = Vec::new();
-    flatten_use_path(
-      &use_path.tree,
-      vec![syn::PathSegment {
-        ident: use_path.ident.clone(),
-        arguments: syn::PathArguments::None,
-      }],
-      &mut flat_paths,
-    );
-
-    // Step 2: Build a prefix tree (trie) from the flat paths
-    #[derive(Debug, Default)]
-    struct TrieNode {
-      name: String,
-      children: Vec<TrieNode>,
-    }
-
-    impl TrieNode {
-      fn is_leaf(&self) -> bool {
-        self.children.is_empty()
-      }
-    }
-
-    fn insert_path(node: &mut TrieNode, segments: &[syn::PathSegment]) {
-      if segments.is_empty() {
-        return;
-      }
-      let ident = segments[0].ident.to_string();
-      if let Some(mut child) = node.children.iter_mut().find(|child| child.name == ident) {
-        insert_path(&mut child, &segments[1..])
-      } else {
-        let mut new_child = TrieNode {
-          name: ident.clone(),
-          children: vec![],
-        };
-        insert_path(&mut new_child, &segments[1..]);
-        node.children.push(new_child);
-      }
-    }
-
-    let mut root = TrieNode {
-      name: "__root__".to_string(),
-      ..Default::default()
-    };
-
-    for path in &flat_paths {
-      let path = self.types.get(path).unwrap_or(path);
-      match path {
-        syn::Type::Path(tp) => {
-          let segments: Vec<_> = tp.path.segments.iter().cloned().collect();
-          insert_path(&mut root, &segments);
-        }
-        _ => {
-          panic!("Expected a Type::Path, got: {:?}", path)
-        }
-      }
-    }
-
-    // Step 3: Recursively convert the trie to UseTree
-    fn trie_to_use_tree(node: &TrieNode) -> syn::UseTree {
-      if node.is_leaf() {
-        let ident = syn::Ident::new(&node.name, proc_macro2::Span::call_site());
-        if &node.name == GLOB_IDENTIFIER {
-          syn::UseTree::Glob(syn::UseGlob {
-            star_token: Default::default(),
-          })
-        } else {
-          syn::UseTree::Name(syn::UseName { ident })
-        }
-      } else if node.children.len() == 1 {
-        // single child, add a path for it
-        let ident = syn::Ident::new(&node.name, proc_macro2::Span::call_site());
-        syn::UseTree::Path(syn::UsePath {
-          ident,
-          colon2_token: Default::default(),
-          tree: Box::new(trie_to_use_tree(&node.children[0])),
-        })
-      } else {
-        // Multiple children, group them
-        let items = node.children.iter().map(trie_to_use_tree).collect();
-        let group = syn::UseTree::Group(syn::UseGroup {
-          brace_token: Default::default(),
-          items,
-        });
-        let ident = syn::Ident::new(&node.name, proc_macro2::Span::call_site());
-        syn::UseTree::Path(syn::UsePath {
-          ident,
-          colon2_token: Default::default(),
-          tree: Box::new(group),
-        })
-      }
-    }
-
-    // The root ident and colon2_token come from the original use_path
-    let rebuilt_tree = trie_to_use_tree(&root);
+    let rebuilt_tree = self.rebuild_use_path(&use_path);
 
     match rebuilt_tree {
       UseTree::Path(UsePath {
