+++ synca/src/fold/mod.rs
@@ -1,5 +1,6 @@
 mod attrs;
 mod fold;
+mod use_path;
 
 pub use attrs::*;
 pub use fold::*;
